Caching setup
This includes a (currently breaking, since not functioning) test to
expose that a fresh call has the same output as a cached call.

This has some duplications; as this is very much a work in progress.

Todo: Figure out where the result from the call to `@context.render`
(in the methods `_render_partial` and `_render_active_model_partial`)
ends up, so I can _actually_ cache it 😬