Replace enum messages, don't concut them (#55)
# Double definition bug
```
pb.messages "This is the first string of the repeating ones"
pb.messages "This is the second one"
```
This code right now equals to this:

`pb.messages ["This is the first string of the repeating ones", "This is the second one"] `

But jbuilder handles this case differently. So in this PR behaviour is change to one similar to jbuilder - only last definition will be persisted and first one will be discarded.

`pb.messages ["This is the second one"]`

# Double render bug
It seems that rails-twirp gem does double rendering sometimes and some fields have duplicated values. e.g.

`pb.redemption_methods [:instore, :online]`

could end up in putput as `[:instore, :online, :instore, :online]`