When setting repeated fields, replace them instead of appending (#55)
This is a breaking change. Previously, we would handle repeated fields
as something you can call multiple times on the builder, like so:

```
pb.messages "This is the first string of the repeating ones"
pb.messages "This is the second one"
```

From now, this is no longer going to work and you need to assign all the messages
in one go:

`pb.messages ["This is the first string of the repeating ones", "This is the second one"] `

This behaviour is in alignment with how `jbuilder` handles this. The last assignment wins, the
previous one will be discarded.

`pb.messages ["This is the second one"]`

It seems that rails-twirp gem does double rendering sometimes and some fields have duplicated values. e.g.

`pb.redemption_methods [:instore, :online]`

could end up in putput as `[:instore, :online, :instore, :online]`

That bug we are going to fix as well, but for now making repeated fields safer
is a good first step.